Moves FormFieldText
