+++ src/components/Field/index.tsx
@@ -49,3 +49,6 @@ export * from "./FormFieldRaw";
 
 export { default as FormFieldTable } from "./FormFieldTable";
 export * from "./FormFieldTable";
+
+export { default as FormFieldText } from "./FormFieldText";
+export * from "./FormFieldText";
